@@ -33,7 +33,6 @@ version = find_version('luma', 'emulator', '__init__.py')
 needs_pytest = {'pytest', 'test', 'ptr'}.intersection(sys.argv)
 pytest_runner = ['pytest-runner'] if needs_pytest else []
 test_deps = [
-    'mock;python_version<"3.3"',
     'pytest<=4.5',
     'pytest-cov'
 ]
@@ -59,7 +58,7 @@ setup(
         "luma/emulator/images/7-segment.png"
     ]},
     install_requires=[
-        "luma.core>=1.12.0",
+        "luma.core>=1.14.0",
         "pygame"
     ],
     setup_requires=pytest_runner,
@@ -75,7 +74,7 @@ setup(
         'test': test_deps
     },
     zip_safe=False,
-    python_requires='>=2.7',
+    python_requires='>=3.5, <4',
     classifiers=[
         "License :: OSI Approved :: MIT License",
         "Development Status :: 5 - Production/Stable",
@@ -85,12 +84,10 @@ setup(
         "Topic :: System :: Hardware",
         "Topic :: System :: Hardware :: Hardware Drivers",
         "Topic :: Software Development :: Libraries :: pygame",
-        "Programming Language :: Python :: 2",
-        "Programming Language :: Python :: 2.7",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.4",
         "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7"
+        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8"
     ]
 )
